mcp-examples/09-iac-analyzer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iac-analyzer-mcp"
version = "0.1.0"
description = "Infrastructure as Code Analyzer MCP Server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "python-hcl2>=4.3.2",
    "checkov>=3.1.9",
    "cfn-lint>=0.83.3",
    "ansible-lint>=6.22.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.2", "black>=23.9.1", "ruff>=0.0.292"]
